backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pandas
numpy
pillow
pypdf
matplotlib
requests
